Separate file for `RStateDelegate`
